refactor(frontend): consolidate UI primitives and remove index exports
