--- frontend/src/components/ui/SectionTitle.tsx
@@ -1,32 +1,0 @@
-import { cn } from "@/lib/utils";
-
-type SectionTitleProps = React.HTMLAttributes<HTMLHeadingElement> & {
-	size?: "sm" | "md" | "lg" | "xl";
-	as?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
-};
-
-export function SectionTitle({
-	className,
-	size = "md",
-	as = "h2",
-	children,
-	...props
-}: SectionTitleProps) {
-	const sizeClasses = {
-		sm: "text-sm font-semibold",
-		md: "text-lg font-bold",
-		lg: "text-xl font-bold",
-		xl: "text-3xl font-bold",
-	};
-
-	const Tag = as;
-
-	return (
-		<Tag
-			className={cn("mb-2 text-gray-900", sizeClasses[size], className)}
-			{...props}
-		>
-			{children}
-		</Tag>
-	);
-}
