--- frontend/src/components/ui/Skeleton.tsx
@@ -1,12 +1,0 @@
-import { cn } from "@/lib/utils";
-
-interface SkeletonProps extends React.HTMLAttributes<HTMLDivElement> {}
-
-export function Skeleton({ className, ...props }: SkeletonProps) {
-	return (
-		<div
-			className={cn("animate-pulse rounded-md bg-gray-200", className)}
-			{...props}
-		/>
-	);
-}
