--- frontend/src/components/ui/index.ts
@@ -1,5 +1,0 @@
-export { Button } from "./Button";
-export { Card, KPICard } from "./Cards";
-export { ColorPicker, FormError, Input, Select } from "./FormControls";
-export { SectionTitle } from "./SectionTitle";
-export { Skeleton } from "./Skeleton";
